+++ CHANGELOG.md
@@ -2,8 +2,7 @@
 
 ## 1.7.5 - Unreleased
 ### Changed
-- User actions are now logged in `storage/logs/campaign.log` file.
-- Failures from imports now get logged in `storage/logs/campaign.log` file.
+- User actions, import failures and some exceptions are now logged in the `storage/logs/campaign.log` file.
 
 ### Fixed
 - Fixed pending sendout count not displaying in pending scheduled sendouts.